refactor: arranged functions differently
